pyproject.toml
[project]
name = "faaster"
version = "1.0.0-alpha"
description = "Framework for automated deployment of Reactive Asset Administration Shell (Type 2) over OPC UA"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
authors = [
    { name = "Alison Soares de Almeida", email = "alison.almeida@ufam.edu.br" }
]
keywords = [
    "aas",
    "asset-administration-shell",
    "opc-ua",
    "industry-4.0",
    "digital-twin",
    "iiot",
    "rami40",
    "idta"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Manufacturing",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
system = [
    "uvloop (>=0.22.1,<0.23.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "asyncua (>=1.1.8,<2.0.0)",
    "pydantic (>=2.13.1,<3.0.0)",
    "gmqtt (>=0.7.0,<0.8.0)",
    "aioamqp (>=0.15.0,<0.16.0)",
    "pymongo (>=4.16.0,<5.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "pycountry (>=26.2.16,<27.0.0)",
    "basyx-python-sdk (>=2.0.0,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)"
]
